Move utility class to Glue v1 package
